refactor: 디렉토리 네임 변경
constant -> constants 변경
